chore : domain 멀티 모듈 core로 통합
