--- domain/my-pet-auth-domain/build.gradle
@@ -1,11 +1,0 @@
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-
-    // jwt
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-
-    implementation project(':my-pet-common')
-    implementation project(':domain:my-pet-member-domain')
-}
