--- domain/my-pet-auth-domain/src/main/java/com/shshon/mypet/auth/infra/JwtTokenProvider.java
@@ -1,59 +1,0 @@
-package com.shshon.mypet.auth.infra;
-
-import com.shshon.mypet.auth.properties.SecurityProperties;
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-
-@Component
-public class JwtTokenProvider {
-
-    private final Key secretKey;
-    private final Long validityInMilliseconds;
-
-    public JwtTokenProvider(SecurityProperties securityProperties) {
-        SecurityProperties.Token tokenProperty = securityProperties.getJwt().getToken();
-        String secretKey = tokenProperty.getSecretKey();
-
-        this.secretKey = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-        this.validityInMilliseconds = tokenProperty.getValidityInMilliseconds();
-    }
-
-    public String createToken(String email) {
-        Claims claims = Jwts.claims().setSubject(email);
-        Date now = new Date();
-        Date validity = new Date(now.getTime() + validityInMilliseconds);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(validity)
-                .signWith(secretKey, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            parseClaims(token)
-                    .getExpiration()
-                    .before(new Date());
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    private Claims parseClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(secretKey)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-}
