--- domain/my-pet-member-domain/build.gradle
@@ -1,6 +1,0 @@
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-
-    implementation project(":my-pet-common")
-    implementation project(":my-pet-infra")
-}
