+++ my-pet-api/build.gradle
@@ -15,16 +15,13 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
-    // sub modules
-    implementation project(':my-pet-common')
-    implementation project(':my-pet-infra')
-    implementation project(':domain:my-pet-auth-domain')
-    implementation project(':domain:my-pet-member-domain')
-
-    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
     // db
     implementation 'mysql:mysql-connector-java'
-    testImplementation 'com.h2database:h2'
+
+    // sub modules
+    implementation project(':my-pet-core')
+
+    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
 
     testImplementation 'io.rest-assured:rest-assured:5.3.0'
     asciidoctorExt 'org.springframework.restdocs:spring-restdocs-asciidoctor'
