+++ my-pet-api/src/main/java/com/shshon/mypet/MyPetApplication.java
@@ -1,12 +1,18 @@
 package com.shshon.mypet;
 
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
 
 @SpringBootApplication
-public class MyPetApplication {
+public class MyPetApplication implements CoreApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(MyPetApplication.class, args);
+        new MyPetApplication().start(args);
+    }
+
+    @Override
+    public void start(String[] args) {
+        SpringApplicationBuilder springApplicationBuilder = new SpringApplicationBuilder(MyPetApplication.class);
+        springApplicationBuilder.run(args);
     }
 }
