+++ my-pet-api/src/main/java/com/shshon/mypet/endpoint/v1/auth/request/LoginMemberRequest.java
@@ -2,8 +2,6 @@ package com.shshon.mypet.endpoint.v1.auth.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
-import com.shshon.mypet.member.domain.Member;
-import com.shshon.mypet.member.dto.MemberDto;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
