+++ my-pet-core/src/main/java/com/shshon/mypet/CoreApplication.java
@@ -1,0 +1,6 @@
+package com.shshon.mypet;
+
+public interface CoreApplication {
+
+    void start(String[] args);
+}
