+++ my-pet-core/src/main/java/com/shshon/mypet/auth/service/AuthService.java
