+++ my-pet-core/src/main/java/com/shshon/mypet/common/exception/MpException.java
