+++ my-pet-core/src/main/java/com/shshon/mypet/common/exception/ServerException.java
