+++ my-pet-core/src/main/java/com/shshon/mypet/member/domain/MemberCertification.java
