+++ my-pet-core/src/main/java/com/shshon/mypet/member/exception/AlreadyCertificatedMemberException.java
