+++ my-pet-core/src/main/java/com/shshon/mypet/member/exception/AuthorizationException.java
