+++ my-pet-core/src/main/resources/static/docs/index.html
@@ -584,7 +584,7 @@ Location: v1/member/1</code></pre>
 <h4 id="Member-인증_curl_request"><a class="link" href="#Member-인증_curl_request">Curl request</a></h4>
 <div class="listingblock">
 <div class="content">
-<pre class="highlightjs highlight"><code class="language-bash hljs" data-lang="bash">$ curl 'http://localhost:8080/v1/member/certification?code=0f060a01-bb57-40ea-998b-0ed01fd5b41c' -i -X GET \
+<pre class="highlightjs highlight"><code class="language-bash hljs" data-lang="bash">$ curl 'http://localhost:8080/v1/member/certification?code=cce70e78-a1e0-4b90-ac2a-028133882769' -i -X GET \
     -H 'Content-Type: application/json;charset=UTF-8' \
     -H 'Accept: application/json'</code></pre>
 </div>
@@ -594,7 +594,7 @@ Location: v1/member/1</code></pre>
 <h4 id="Member-인증_http_request"><a class="link" href="#Member-인증_http_request">HTTP request</a></h4>
 <div class="listingblock">
 <div class="content">
-<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">GET /v1/member/certification?code=0f060a01-bb57-40ea-998b-0ed01fd5b41c HTTP/1.1
+<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">GET /v1/member/certification?code=cce70e78-a1e0-4b90-ac2a-028133882769 HTTP/1.1
 Content-Type: application/json;charset=UTF-8
 Accept: application/json
 Host: localhost:8080</code></pre>
@@ -750,7 +750,7 @@ Content-Length: 159
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-02-16 21:38:29 +0900
+Last updated 2023-02-16 18:27:00 +0900
 </div>
 </div>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.18.3/highlight.min.js"></script>
