+++ my-pet-core/src/test/java/com/shshon/mypet/pet/service/CreatePetServiceTest.java
@@ -1,0 +1,12 @@
+package com.shshon.mypet.pet.service;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class CreatePetServiceTest {
+
+    @Test
+    @DisplayName("회원 반려동물 등록 요청시 등록 후 반려동물 정보를 반환한다")
+    void createMyPetTest() {
+    }
+}
