+++ my-pet-infra/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
+    // spring
     implementation 'org.springframework.boot:spring-boot-starter-mail'
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
 }
