+++ settings.gradle
@@ -1,6 +1,4 @@
 rootProject.name = 'my-pet'
+include("my-pet-core")
 include("my-pet-api")
-include("my-pet-common")
 include("my-pet-infra")
-include("domain:my-pet-member-domain")
-include("domain:my-pet-auth-domain")
